The KBS guidelines are unclear as the coding value of Medication under certain situations. (#323)
This change updates the coding such that

1. If rxnorm information is available, return it as the code.

2. If rxnorm is not available, but product is, then this is a "local" drug. We will use the
local drug name as the text and the code.coding.display value. The product.id is the VA
specific "VUID" medication ID. We want to use this value as the code.coding.code along with the
VA specific system.

3. If neither rxnorm or product is available, we'll create a code with no coding, using just
the local drug name as text.